network/retrieval: pr comments
